Galleon : Test : Committing failing test.check defspec
Committing a failing defspec to show basic test.check test and what
the failing report looks like.

This first test was to confirm that converting EDN to JSON and back
resulted in identical data structures. Apparrently, that is not
a valid assumption as the test fails (have not yet investigated).

--Tom